[level 1] Title: 자연수 뒤집어 배열로 만들기, Time: 0.05 ms, Memory: 88.6 MB -BaekjoonHub

+++ 프로그래머스/1/12932. 자연수 뒤집어 배열로 만들기/자연수 뒤집어 배열로 만들기.java
@@ -1,12 +1,12 @@
 import java.util.*;
 
 class Solution {
-    public List<Long> solution(long n) {
-        List<Long> answer = new ArrayList<>();
-        while (n > 0) {
-            answer.add(n % 10);
-            n /= 10;
-        }
-        return answer;
-    }
-}
+     public List<Integer> solution(long n) {
+         List<Integer> answer = new ArrayList<>();
+         while (n > 0) {
+             answer.add((int) (n % 10));
+             n /= 10;
+         }
+         return answer;
+     }
+ }
